Move FlutterServiceApp to separate file
