+++ embedding/cpp/flutter_app.cc
@@ -193,92 +193,3 @@ void FlutterApp::ParseEngineArgs() {
     TizenLog::Warn("Error removing file: %s", strerror(errno));
   }
 }
-
-int FlutterServiceApp::Run(int argc, char **argv) {
-  service_app_lifecycle_callback_s lifecycle_cb = {};
-  lifecycle_cb.create = [](void *data) -> bool {
-    FlutterApp *app = (FlutterApp *)data;
-    return app->OnCreate();
-  };
-  lifecycle_cb.terminate = [](void *data) {
-    FlutterApp *app = (FlutterApp *)data;
-    app->OnTerminate();
-  };
-  lifecycle_cb.app_control = [](app_control_h a, void *data) {
-    FlutterApp *app = (FlutterApp *)data;
-    app->OnAppControlReceived(a);
-  };
-
-  app_event_handler_h handler;
-  service_app_add_event_handler(&handler, APP_EVENT_LOW_MEMORY,
-                                [](app_event_info_h e, void *data) {
-                                  FlutterApp *app = (FlutterApp *)data;
-                                  app->OnLowMemory(e);
-                                },
-                                this);
-  service_app_add_event_handler(&handler, APP_EVENT_LOW_BATTERY,
-                                [](app_event_info_h e, void *data) {
-                                  FlutterApp *app = (FlutterApp *)data;
-                                  app->OnLowBattery(e);
-                                },
-                                this);
-  service_app_add_event_handler(&handler, APP_EVENT_LANGUAGE_CHANGED,
-                                [](app_event_info_h e, void *data) {
-                                  FlutterApp *app = (FlutterApp *)data;
-                                  app->OnLanguageChanged(e);
-                                },
-                                this);
-  service_app_add_event_handler(&handler, APP_EVENT_REGION_FORMAT_CHANGED,
-                                [](app_event_info_h e, void *data) {
-                                  FlutterApp *app = (FlutterApp *)data;
-                                  app->OnRegionFormatChanged(e);
-                                },
-                                this);
-
-  int ret = service_app_main(argc, argv, &lifecycle_cb, this);
-  if (ret != APP_ERROR_NONE) {
-    TizenLog::Error("Could not launch a Service application. (%d)", ret);
-  }
-  return ret;
-}
-
-bool FlutterServiceApp::OnCreate() {
-  TizenLog::Debug("Launching a Flutter Service application...");
-
-  std::string res_path;
-  {
-    auto path = app_get_resource_path();
-    if (path == nullptr) {
-      TizenLog::Error("Could not obtain the app directory info.");
-      return false;
-    }
-    res_path = path;
-    free(path);
-  }
-  std::string assets_path(res_path + "/flutter_assets");
-  std::string icu_data_path(res_path + "/icudtl.dat");
-  std::string aot_lib_path(res_path + "/../lib/libapp.so");
-
-  // Read engine arguments passed from the tool.
-  ParseEngineArgs();
-
-  std::vector<const char *> switches;
-  for (auto &arg : engine_args) {
-    switches.push_back(arg.c_str());
-  }
-
-  FlutterDesktopEngineProperties engine_prop = {};
-  engine_prop.assets_path = assets_path.c_str();
-  engine_prop.icu_data_path = icu_data_path.c_str();
-  engine_prop.aot_library_path = aot_lib_path.c_str();
-  engine_prop.switches = switches.data();
-  engine_prop.switches_count = switches.size();
-
-  handle = FlutterDesktopRunEngine(engine_prop, false);
-  if (!handle) {
-    TizenLog::Error("Could not launch a Flutter Service application.");
-    return false;
-  }
-
-  return true;
-}
