+++ embedding/cpp/include/flutter_service_app.h
@@ -1,0 +1,25 @@
+// Copyright 2021 Samsung Electronics Co., Ltd. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#ifndef FLUTTER_TIZEN_EMBEDDING_CPP_INCLUDE_FLUTTER_SERVICE_APP_H_
+#define FLUTTER_TIZEN_EMBEDDING_CPP_INCLUDE_FLUTTER_SERVICE_APP_H_
+
+#include <flutter/plugin_registry.h>
+#include <flutter_tizen.h>
+#include <service_app.h>
+
+#include "flutter_app.h"
+
+// The app base class for headless execution.
+class FlutterServiceApp : public FlutterApp {
+ public:
+  explicit FlutterServiceApp() {}
+  ~FlutterServiceApp() {}
+
+  bool OnCreate() override;
+
+  int Run(int argc, char **argv) override;
+};
+
+#endif /* FLUTTER_TIZEN_EMBEDDING_CPP_INCLUDE_FLUTTER_SERVICE_APP_H_ */
